Fixed exporter failure in presence of empty tags.
It is possible that Kannel is configured with some parameters left empty.  This
causes xmltodict.parse return a result with values for these keys of NoneType.
As a result this leads to crash inside Prometheus client library (exposition.py)
when it attempts to escape some symbols in dict values.

This patch fixes the problem by post-processing the resulting dict from
xmltodict.parse with a function that replaces all None values with empty
strings.

Additionally replaced "sms-id" label name with "sms_id" since Prometheus 2.0
does not allow dashes in label names.